+++ src/transformers/models/groupvit/modeling_tf_groupvit.py
@@ -1634,7 +1634,7 @@ class TFGroupViTTextModel(TFGroupViTPreTrainedModel):
             `Dict[str, tf.Tensor]`: The dummy inputs.
         """
         return {
-            "input_ids": tf.constant(DUMMY_INPUTS, dtype=tf.int32),
+            "input_ids": tf.constant(DUMMY_INPUTS, dtype=tf.int64),
         }
 
     @tf.function(
